Save and load estimator, export feature_importance

+++ Predict_tournament/train_model.py
@@ -58,22 +58,29 @@ def load_from_csv(path, delimiter=','):
     return pd.read_csv(path, delimiter=delimiter,dtype=float)
 
 def create_estimator():
-    # Loading data
-    prefix = '/home/tom/Documents/Master1_DataScience/1er QUADRI/Big-Data-Project/Data_cleaning/'
-    df = load_from_csv(os.path.join(prefix, 'merged_matches_players.csv'))
 
-    y = df['PlayerA Win'].values.squeeze()
-    train_features = df.drop(columns=['PlayerA Win']).columns
-    X = df.drop(columns=['PlayerA Win']).values.squeeze()  
+    model = None
+    filename = "estimators/RandomForest_depth3.pkl"
+    if(os.path.isfile(filename)):
+        model = joblib.load(filename)
+    else:
+        # Loading data
+        prefix = '/home/tom/Documents/Master1_DataScience/1er QUADRI/Big-Data-Project/Data_cleaning/'
+        df = load_from_csv(os.path.join(prefix, 'merged_matches_players.csv'))
 
-    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
+        y = df['PlayerA Win'].values.squeeze()
+        train_features = df.drop(columns=['PlayerA Win']).columns
+        X = df.drop(columns=['PlayerA Win']).values.squeeze()  
 
-    print("Training...getting most important features")
-
-    with measure_time('Training...getting most important features'):
-        model = RandomForestClassifier(n_estimators=1000,max_depth=3, bootstrap=True,n_jobs=-1, random_state=42)
-        model.fit(X_train,y_train)
-        joblib.dump(model, "estimators/RandomForest_depth3.pkl") 
+        X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
+        
+        #Training
+        print("Training... getting most important features")
+        
+        with measure_time('Training...getting most important features'):
+            model = RandomForestClassifier(n_estimators=1000,max_depth=3, bootstrap=True,n_jobs=-1, random_state=42)
+            model.fit(X_train,y_train)
+            joblib.dump(model, "estimators/RandomForest_depth3.pkl") 
 
     feature_importances = pd.DataFrame(model.feature_importances_,
                                         index = train_features,
@@ -81,7 +88,7 @@ def create_estimator():
 
     print("Most important features")
     print(feature_importances[:100])
-    most_imp_features = feature_importances[:100].axes[0].tolist()
+    feature_importances[:100].to_csv("feature_importance.csv")
     
     y_pred = model.predict(X_test)
     print("Test set accuracy: {}".format(accuracy_score(y_test, y_pred)))
